Estimate step or beta selection model

Estimate step or beta selection model, with standard errors and confidence intervals based on either cluster-robust variance estimators (i.e., sandwich estimators) or cluster-level bootstrapping to handle dependent effect size estimates.

Usage

selection_model(
  data,
  yi,
  sei,
  pi,
  ai,
  cluster,
  selection_type = c("step", "beta"),
  steps = NULL,
  mean_mods = NULL,
  var_mods = NULL,
  sel_mods = NULL,
  sel_zero_mods = NULL,
  subset = NULL,
  estimator = "ML",
  vcov_type = "robust",
  CI_type = "large-sample",
  conf_level = 0.95,
  theta = NULL,
  optimizer = NULL,
  optimizer_control = list(),
  use_jac = TRUE,
  bootstrap = "none",
  R = 1999,
  ...
)

Arguments

data

data.frame or tibble containing the meta-analytic data

yi

vector of effect sizes estimates.

sei

vector of sampling standard errors.

pi

optional vector of one-sided p-values. If not specified, p-values will be computed from yi and sei.

ai

optional vector of analytic weights.

cluster

vector indicating which observations belong to the same cluster.

selection_type

character string specifying the type selection model to estimate, with possible options "step" or "beta".

steps

If selection_type = "step", a numeric vector of one or more values specifying the thresholds (or steps) where the selection probability changes, with a default of steps = .025. If selection_type = "beta", then a numeric vector of two values specifying the thresholds beyond which the selection function is truncated, with a default of steps = c(.025, .975).

mean_mods

optional model formula for moderators related to average effect size magnitude.

var_mods

optional model formula for moderators related to effect size heterogeneity.

sel_mods

optional model formula for moderators related to the probability of selection. Only relevant for selection_type = "step".

sel_zero_mods

optional model formula for moderators related to the probability of selection for p-values below the lowest threshold value of steps. Only relevant for selection_type = "step".

subset

optional logical expression indicating a subset of observations to use for estimation.

estimator

vector indicating whether to use the maximum likelihood or the hybrid estimator, with possible options "ML", "hybrid", and "hybrid-full". If selection_type = "beta", only the maximum likelihood estimator, "ML", is available. For step function models, both maximum likelihood and hybrid estimators are available.

vcov_type

character string specifying the type of variance-covariance matrix to calculate, with possible options "robust" for robust or cluster-robust standard errors, "model-based" for model-based standard errors, or "none".

CI_type

character string specifying the type of confidence interval to calculate, with possible options "large-sample" for large-sample normal (the default), "percentile" for percentile bootstrap, "basic" for basic bootstrap, "student" for studentized bootstrap, or "none".

conf_level

desired coverage level for confidence intervals, with the default value set to .95

theta

optional numeric vector of starting values to use in optimization routines.

optimizer

character string indicating the optimizer to use. Ignored if estimator = "hybrid" or "hybrid-full".

optimizer_control

an optional list of control parameters to be used for optimization

use_jac

logical with TRUE (the default) indicating to use the Jacobian of the estimating equations for optimization.

bootstrap

character string specifying the type of bootstrap to run, with possible options "none" (the default), "exponential" for the fractionally re-weighted cluster bootstrap, or "multinomial" for a conventional clustered bootstrap.

R

number of bootstrap replications, with a default of 1999.

...

further arguments passed to simhelpers::bootstrap_CIs.

Value

A numeric vector.

Examples

res_ML <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = 0.025,
  estimator = "ML",
  bootstrap = "none"
)

res_ML
#>     param    Est     SE   CI_lo CI_hi
#>      beta 0.2200 0.0538 0.11443 0.325
#>      tau2 0.0395 0.0297 0.00902 0.173
#>  lambda_1 1.0386 0.5188 0.39020 2.765
summary(res_ML)
#> Step Function Model 
#>  
#> Call: 
#> selection_model(data = self_control, yi = g, sei = se_g, cluster = studyid, 
#>     steps = 0.025, estimator = "ML", bootstrap = "none")
#> 
#> Number of clusters = 33; Number of effects = 158
#> 
#> Steps: 0.025 
#> Estimator: maximum likelihood 
#> Variance estimator: robust 
#> 
#> Log composite likelihood of selection model: -54.97399
#> Inverse selection weighted partial log likelihood: 86.79868 
#> 
#> Mean effect estimates:                                       
#>                            Large Sample
#>  Coef. Estimate Std. Error Lower  Upper
#>   beta     0.22     0.0538 0.114  0.325
#> 
#> Heterogeneity estimates:                                         
#>                              Large Sample
#>  Coef. Estimate Std. Error   Lower  Upper
#>   tau2   0.0395     0.0297 0.00902  0.173
#> 
#> Selection process estimates:                                          
#>                               Large Sample
#>     Coef. Estimate Std. Error Lower  Upper
#>  lambda_1     1.04      0.519  0.39   2.76

# configure progress bar
progressr::handlers(global = TRUE)
#> Error in globalCallingHandlers(condition = global_progression_handler): should not be called with handlers on the stack

res_hybrid <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = 0.025,
  estimator = "hybrid",
  bootstrap = "multinomial",
  CI_type = "percentile",
  R = 19
)

res_hybrid
#>     param    Est     SE percentile_lower percentile_upper
#>      beta 0.2197 0.0484          0.14291           0.3291
#>      tau2 0.0394 0.0303          0.00124           0.0937
#>  lambda_1 1.0366 0.1496          0.18127           4.8121
summary(res_hybrid)
#> Step Function Model with Cluster Bootstrapping 
#>  
#> Call: 
#> selection_model(data = self_control, yi = g, sei = se_g, cluster = studyid, 
#>     steps = 0.025, estimator = "hybrid", CI_type = "percentile", 
#>     bootstrap = "multinomial", R = 19)
#> 
#> Number of clusters = 33; Number of effects = 158
#> 
#> Steps: 0.025 
#> Estimator: hybrid estimating equations 
#> Variance estimator: robust 
#> Bootstrap type: multinomial 
#> Number of bootstrap replications: 19 
#> 
#> Log composite likelihood of selection model: -54.97401
#> Inverse selection weighted partial log likelihood: 86.96903 
#> 
#> Mean effect estimates:                                               
#>                            Percentile Bootstrap
#>  Coef. Estimate Std. Error      Lower     Upper
#>   beta     0.22     0.0484      0.143     0.329
#> 
#> Heterogeneity estimates:                                               
#>                            Percentile Bootstrap
#>  Coef. Estimate Std. Error      Lower     Upper
#>   tau2   0.0394     0.0303    0.00124    0.0937
#> 
#> Selection process estimates:                                                  
#>                               Percentile Bootstrap
#>     Coef. Estimate Std. Error      Lower     Upper
#>  lambda_1     1.04       0.15      0.181      4.81