Calculate model-implied weights for specified p-values.

Calculates the selection weights implied by an estimated model for a user-specified p-value or set of p-values.

Usage

selection_wts(mod, pvals, ...)

# S3 method for class 'step.selmodel'
selection_wts(mod, pvals, bootstraps = TRUE, ...)

# S3 method for class 'beta.selmodel'
selection_wts(mod, pvals, bootstraps = TRUE, ...)

Arguments

mod

Fitted model of class "selmodel".

pvals

Numeric vector of p-values for which to calculate selection weights.

...

further arguments passed to some methods.

bootstraps

If mod includes bootstrap replications, then setting bootstraps = TRUE will return selection weights for each bootstrap replication, in addition to the selection weights implied by the model parameter estimates. Ignored if mod does not include bootstrap replications.

Value

If mod does not include bootstrapped confidence intervals or if the argument bootstraps = FALSE, then selection_wts will return a data.frame containing the user-specified p-values and the selection weights implied by the estimated model parameters.

If mod does include bootstrapped confidence intervals (i.e., when inherits(mod, "boot.selmodel") is TRUE) and the argument bootstraps = TRUE, then selection_wts will return a list with two elements. The first element is a data.frame containing the user-specified p-values and the selection weights implied by the estimated model parameters. The second element is a data.frame containing the user-specified p-values and the selection weights implied by each bootstrap replicate of the model parameter estimates. The data.frame includes an additional variable, rep, identifying the bootstrap replicate.

Examples

mod <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = c(0.025, .5),
  estimator = "ML",
  bootstrap = "none"
)

selection_wts(mod, pvals = seq(0, 1, 0.2))
#>     p        wt
#> 1 0.0 1.0000000
#> 2 0.2 0.8555420
#> 3 0.4 0.8555420
#> 4 0.6 0.3614923
#> 5 0.8 0.3614923
#> 6 1.0 0.3614923

mod_boot <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = c(0.025, .5),
  estimator = "ML",
  bootstrap = "multinomial",
  CI_type = "percentile",
  R = 9
)

selection_wts(mod_boot, pvals = seq(0, 1, 0.2))
#> $wts
#>     p        wt
#> 1 0.0 1.0000000
#> 2 0.2 0.8555420
#> 3 0.4 0.8555420
#> 4 0.6 0.3614923
#> 5 0.8 0.3614923
#> 6 1.0 0.3614923
#> 
#> $boot_wts
#>    rep   p        wt
#> 1    1 0.0 1.0000000
#> 2    1 0.2 0.9893249
#> 3    1 0.4 0.9893249
#> 4    1 0.6 0.2897322
#> 5    1 0.8 0.2897322
#> 6    1 1.0 0.2897322
#> 7    2 0.0 1.0000000
#> 8    2 0.2 0.8408823
#> 9    2 0.4 0.8408823
#> 10   2 0.6 0.3392195
#> 11   2 0.8 0.3392195
#> 12   2 1.0 0.3392195
#> 13   3 0.0 1.0000000
#> 14   3 0.2 1.2379700
#> 15   3 0.4 1.2379700
#> 16   3 0.6 0.7158056
#> 17   3 0.8 0.7158056
#> 18   3 1.0 0.7158056
#> 19   4 0.0 1.0000000
#> 20   4 0.2 0.7109054
#> 21   4 0.4 0.7109054
#> 22   4 0.6 0.3367876
#> 23   4 0.8 0.3367876
#> 24   4 1.0 0.3367876
#> 25   5 0.0 1.0000000
#> 26   5 0.2 0.5776662
#> 27   5 0.4 0.5776662
#> 28   5 0.6 0.1866829
#> 29   5 0.8 0.1866829
#> 30   5 1.0 0.1866829
#> 31   6 0.0 1.0000000
#> 32   6 0.2 0.5696899
#> 33   6 0.4 0.5696899
#> 34   6 0.6 0.2514365
#> 35   6 0.8 0.2514365
#> 36   6 1.0 0.2514365
#> 37   7 0.0 1.0000000
#> 38   7 0.2 0.7180944
#> 39   7 0.4 0.7180944
#> 40   7 0.6 0.2427214
#> 41   7 0.8 0.2427214
#> 42   7 1.0 0.2427214
#> 43   8 0.0 1.0000000
#> 44   8 0.2 0.8126387
#> 45   8 0.4 0.8126387
#> 46   8 0.6 0.9759344
#> 47   8 0.8 0.9759344
#> 48   8 1.0 0.9759344
#> 49   9 0.0 1.0000000
#> 50   9 0.2 1.0135422
#> 51   9 0.4 1.0135422
#> 52   9 0.6 0.5519481
#> 53   9 0.8 0.5519481
#> 54   9 1.0 0.5519481
#>