Self-Control Training Meta-Analysis

Meta-analytic dataset containing results of primary studies examining the effect of self-control training.

Usage

self_control

Format

A data frame with 158 rows and 26 variables:

studyid

identifier for the study

esid

identifier for the effect size

name

name of the study

g

effect size in form of Hedges' g

var_g

corresponding variance of the effect size

se_g

corresponding standard error of the effect size

outcome

name of the outcome measure

comparison

conditions compared

type_of_treatment

type of treatment/training

length_of_treatment

length of treatment coded in days

publication_status

indicator for whether the study was published

publication_status_new

indicator for whether the study was published (newer version)

publication_year

year that the study was published

research_group

indicator for whether researchers of the study belonged to the strength model group

control_group_quality

quality of the control group with active indicating that the control group worked on some task

gender_ratio

percentage of males in the sample

type_of_outcome

type of outcome

subjectivity_of_outcome_measurement

indicator for the subjectivity of the outcome measure

lab_based_versus_real_world_behavior

indicator for whether the behavior measured was assessed in lab or in the real-world

stamina_versus_strength

indicator for whether the outcome was assessed with (Stamina) or without a preceding effortful task (Strength)

pre_test_measurement

indicator for whether there was a pre and post measure of outcome or only post measure

sample_population

population examined in the study

sample_age

average age of the sample

attrition

percentage of attrition

participant_compensation

type of compensation received by the participants

self_control_potential

indicator for whether the outcome measure required utilization of maximum self-control potential

Source

OSF page for the project

References

Friese, M., Frankenbach, J., Job, V., & Loschelder, D. D. (2017). Does self-control training improve self-control? A meta-analysis. Perspectives on Psychological Science, 12(6), 1077-1099. doi:10.1177/1745691617697076